+++ vms/avm/wallet_service_test.go
@@ -10,77 +10,29 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/ava-labs/avalanchego/api"
-	"github.com/ava-labs/avalanchego/chains/atomic"
-	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/utils/linkedhashmap"
-	"github.com/ava-labs/avalanchego/vms/avm/txs"
-	"github.com/ava-labs/avalanchego/vms/components/keystore"
+	"github.com/ava-labs/avalanchego/snow/engine/common"
+	"github.com/ava-labs/avalanchego/vms/avm/blocks/executor"
 )
-
-// Returns:
-// 1) genesis bytes of vm
-// 2) the VM
-// 3) The wallet service that wraps the VM
-// 4) atomic memory to use in tests
-func setupWS(t *testing.T, isAVAXAsset bool) ([]byte, *VM, *WalletService, *atomic.Memory, *txs.Tx) {
-	var genesisBytes []byte
-	var vm *VM
-	var m *atomic.Memory
-	var genesisTx *txs.Tx
-	if isAVAXAsset {
-		genesisBytes, _, vm, m = GenesisVM(t)
-		genesisTx = GetAVAXTxFromGenesisTest(genesisBytes, t)
-	} else {
-		genesisBytes, _, vm, m = setupTxFeeAssets(t)
-		genesisTx = GetCreateTxFromGenesisTest(t, genesisBytes, feeAssetName)
-	}
-
-	ws := &WalletService{
-		vm:         vm,
-		pendingTxs: linkedhashmap.New[ids.ID, *txs.Tx](),
-	}
-	return genesisBytes, vm, ws, m, genesisTx
-}
-
-// Returns:
-// 1) genesis bytes of vm
-// 2) the VM
-// 3) The wallet service that wraps the VM
-// 4) atomic memory to use in tests
-func setupWSWithKeys(t *testing.T, isAVAXAsset bool) ([]byte, *VM, *WalletService, *atomic.Memory, *txs.Tx) {
-	require := require.New(t)
-
-	genesisBytes, vm, ws, m, tx := setupWS(t, isAVAXAsset)
-
-	// Import the initially funded private keys
-	user, err := keystore.NewUserFromKeystore(ws.vm.ctx.Keystore, username, password)
-	require.NoError(err)
-
-	require.NoError(user.PutKeys(keys...))
-
-	require.NoError(user.Close())
-	return genesisBytes, vm, ws, m, tx
-}
 
 func TestWalletService_SendMultiple(t *testing.T) {
 	require := require.New(t)
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			_, vm, ws, _, genesisTx := setupWSWithKeys(t, tc.avaxAsset)
+			env := setup(t, tc.avaxAsset)
 			defer func() {
-				require.NoError(vm.Shutdown(context.Background()))
-				vm.ctx.Lock.Unlock()
+				require.NoError(env.vm.Shutdown(context.Background()))
+				env.vm.ctx.Lock.Unlock()
 			}()
 
-			assetID := genesisTx.ID()
+			assetID := env.genesisTx.ID()
 			addr := keys[0].PublicKey().Address()
 
-			addrStr, err := vm.FormatLocalAddress(addr)
+			addrStr, err := env.vm.FormatLocalAddress(addr)
 			require.NoError(err)
-			changeAddrStr, err := vm.FormatLocalAddress(testChangeAddr)
+			changeAddrStr, err := env.vm.FormatLocalAddress(testChangeAddr)
 			require.NoError(err)
-			_, fromAddrsStr := sampleAddrs(t, vm, addrs)
+			_, fromAddrsStr := sampleAddrs(t, env.vm, addrs)
 
 			args := &SendMultipleArgs{
 				JSONSpendHeader: api.JSONSpendHeader{
@@ -105,16 +57,23 @@ func TestWalletService_SendMultiple(t *testing.T) {
 				},
 			}
 			reply := &api.JSONTxIDChangeAddr{}
-			vm.timer.Cancel()
-			require.NoError(ws.SendMultiple(nil, args, reply))
+			require.NoError(env.walletService.SendMultiple(nil, args, reply))
 			require.Equal(changeAddrStr, reply.ChangeAddr)
 
-			pendingTxs := vm.txs
-			require.Len(pendingTxs, 1)
+			require.Equal(common.PendingTxs, <-env.issuer)
 
-			require.Equal(pendingTxs[0].ID(), reply.TxID)
+			blkIntf, err := env.vm.BuildBlock(context.Background())
+			require.NoError(err)
 
-			_, err = vm.GetTx(context.Background(), reply.TxID)
+			blk := blkIntf.(*executor.Block)
+			txs := blk.Txs()
+			require.Len(txs, 1)
+			require.Equal(txs[0].ID(), reply.TxID)
+
+			require.NoError(blk.Verify(context.Background()))
+			require.NoError(blk.Accept(context.Background()))
+
+			_, err = env.vm.state.GetTx(reply.TxID)
 			require.NoError(err)
 		})
 	}
